pyproject.toml
[tool.poetry]
name = "aptos-contracts-annotated"
version = "0.1.0"
description = ""
authors = ["Rafal Stozek <rafal.stozek@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
aptos-sdk = "^0.2.0"
PyYAML = "^6.0"
typer = "^0.6.1"
toolz = "^0.12.0"
rich = "^12.5.1"
typing-extensions = "^4.3.0"

[tool.poetry.dev-dependencies]
black = "^22.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
